+++ src/common.cpp
@@ -673,6 +673,13 @@ PyObject *frozenset_from_obj(PyObject *o) {
     return result;
 }
 
+PyObject *memoryview_from_obj(PyObject *o) {
+    PyObject *result = PyMemoryView_FromObject(o);
+    if (!result)
+        raise_python_error();
+    return result;
+}
+
 // ========================================================================
 
 PyObject **seq_get(PyObject *seq, size_t *size_out, PyObject **temp_out) noexcept {